--- src/app/core/components/logo/logo.component.html
@@ -1,3 +1,0 @@
-<a [routerLink]="link">
-  <img ngSrc="./assets/logo.png" width="50" height="50" alt="logo" />
-</a>
