+++ src/app/core/layouts/logo/logo.component.html
@@ -1,0 +1,3 @@
+<a [routerLink]="link">
+  <img ngSrc="../../../../assets/logo.png" width="50" height="50" alt="logo" />
+</a>
