+++ src/app/core/layouts/logo/logo.component.ts
