+++ src/app/core/layouts/menu/menu.component.scss
@@ -1,5 +1,5 @@
 @use "../../../../styles/mixings" as b;
-@import "variables";
+@import "../../../../styles/variables";
 
 .navbar {
   position: absolute;
