+++ src/app/features/contact/contact.component.scss
@@ -1,4 +1,4 @@
-@import "variables";
+@import "../../../styles/variables";
 
 .section-contact {
   display: flex;
