+++ src/app/features/home/home.component.ts
@@ -1,7 +1,7 @@
 import { Component } from '@angular/core';
 import { CommonModule, NgOptimizedImage } from '@angular/common';
-import { LogoComponent } from '@core/components/logo/logo.component';
-import { MenuComponent } from '@core/components/menu/menu.component';
+import { LogoComponent } from '@core/layouts/logo/logo.component';
+import { MenuComponent } from '@core/layouts/menu/menu.component';
 import { Icon } from '@core/models/entities/icon';
 import { Image } from '@core/models/entities/image';
 import { SocialMediaComponent } from '@shared/components/social-media/social-media.component';
