+++ src/app/shared/components/loading/loading.component.html
@@ -1,1 +1,1 @@
-<mat-spinner color="warn"></mat-spinner>
+<mat-spinner color="#0FB5C8"></mat-spinner>
